Rumor Game calculate some probabilty
Rumor Game
A simplified model for the spread of a rumor goes this way: There are
N=20 people in a group of friends, of which some have heard of the
rumor and the others have not. During any single period of time, two
persons are selected at random from the group and assumed to
interact. If one of these persons has heard the rumor and the other
has not, then with a probability 0.15, the rumor is transmitted.
Assuming that this process begins at time t=0 with a single person
knowing the rumor and the first interaction occurs at time t=1, run
the system for 200 times to find the following quantities:
a. Average time that it takes for everyone to hear the rumor.
b. Probability that at least 15 people know the humor at time
t=100.
c. Probability that at most 2 people know the humor at time t=10.
d. What happens to the above quantities in part a, b and c if the
transmission probability increases to 0.20? Are the results
intuitively expected?
e. What happens to the above quantities in part a, b and c if the
transmission probability increases to 0.99? Are the results
intuitively expected?